Align text_buf, take advantage to use and mask &text_buf[R] ptr, not R.
Reorder .bss section and add .align to ensure text_buf has alignment of 2*N,
that is the bottom P_BITS+1 bits of its address are all zero, so an overflow
from the bottom N bits of a pointer into it is harmless. This is zero cost in
executable size.

Convert x2 from representing R to being &text_buf[R]. Allows using
post-increment instead of a separate add.